Created useTransaction hook
