+++ src/hooks/useDeployNonFungibleToken.ts
@@ -3,11 +3,13 @@ import { TxStage } from "../model/Transaction"
 import TransactionsService from "../service/TransactionsService"
 import VIP181Service from "../service/VIP181Service"
 import { getErrorMessage } from "../utils/ExtensionUtils"
+import { useTransaction } from "./useTransaction"
 
 const useDeployNonFungibleToken = () => {
   const [txId, setTxId] = useState<string>()
   const [txStatus, setTxStatus] = useState(TxStage.NONE)
   const [error, setError] = useState<string>()
+  const { requestTransaction } = useTransaction()
 
   const deployNftContract = async (
     accountAddress: string,
@@ -30,7 +32,7 @@ const useDeployNonFungibleToken = () => {
 
       setTxStatus(TxStage.IN_EXTENSION)
 
-      const { txid } = await TransactionsService.requestTransaction(
+      const { txid } = await requestTransaction(
         accountAddress,
         clause,
         comment,
